More strict validation for creating key file.

+++ network_plugin/keypair.py
@@ -23,7 +23,7 @@ from Crypto import Random
 AUTO_GENERATE = 'auto_generate'
 PRIVATE_KEY = 'private_key'
 PUBLIC_KEY = 'public_key'
-CREATE_PRIVATE_KEY_FILE = 'create_private_key_file'
+CREATE_PRIVATE_KEY_FILE = 'create_file'
 PATH = 'path'
 KEY = 'key'
 USER = 'user'
@@ -58,7 +58,7 @@ def create(**kwargs):
         public, private = _generate_pair()
         ctx.instance.runtime_properties[PRIVATE_KEY][KEY] = private
         ctx.instance.runtime_properties[PUBLIC_KEY][KEY] = public
-        if ctx.node.properties.get(CREATE_PRIVATE_KEY_FILE):
+        if ctx.node.properties.get(PRIVATE_KEY, {}).get(CREATE_PRIVATE_KEY_FILE):
             ctx.instance.runtime_properties[PRIVATE_KEY][PATH] = _create_path()
             _save_key_file(ctx.instance.runtime_properties[PRIVATE_KEY][PATH],
                            ctx.instance.runtime_properties[PRIVATE_KEY][KEY])
@@ -67,7 +67,7 @@ def create(**kwargs):
             ctx.node.properties.get(PUBLIC_KEY, {}).get(KEY)
         ctx.instance.runtime_properties[PRIVATE_KEY][KEY] = \
             ctx.node.properties.get(PRIVATE_KEY, {}).get(KEY)
-        if ctx.node.properties.get(CREATE_PRIVATE_KEY_FILE):
+        if ctx.node.properties.get(PRIVATE_KEY, {}).get(CREATE_PRIVATE_KEY_FILE):
             ctx.instance.runtime_properties[PRIVATE_KEY][PATH] = \
                 ctx.node.properties.get(PRIVATE_KEY, {}).get(PATH)
             if ctx.node.properties.get(PRIVATE_KEY, {}).get(KEY):
@@ -79,12 +79,12 @@ def create(**kwargs):
 @operation
 def delete(**kwargs):
     if ctx.node.properties[AUTO_GENERATE]:
-        if ctx.node.properties.get(CREATE_PRIVATE_KEY_FILE):
-            _delete_key_file(ctx.instance.runtime_properties[PRIVATE_KEY][PATH])
+        if ctx.node.properties.get(PRIVATE_KEY, {}).get(CREATE_PRIVATE_KEY_FILE):
+            _delete_key_file(ctx.instance.runtime_properties)
     else:
         if ctx.node.properties.get(PRIVATE_KEY, {}).get(KEY):
-            if ctx.node.properties.get(CREATE_PRIVATE_KEY_FILE):
-                _delete_key_file(ctx.instance.runtime_properties[PRIVATE_KEY][PATH])
+            if ctx.node.properties.get(PRIVATE_KEY, {}).get(CREATE_PRIVATE_KEY_FILE):
+                _delete_key_file(ctx.instance.runtime_properties)
     if PRIVATE_KEY in ctx.instance.runtime_properties:
         del ctx.instance.runtime_properties[PRIVATE_KEY]
     if PUBLIC_KEY in ctx.instance.runtime_properties:
@@ -94,11 +94,12 @@ def delete(**kwargs):
 @operation
 def server_connect_to_keypair(**kwargs):
     host_rt_properties = ctx.source.instance.runtime_properties
+    target_rt_properties = ctx.target.instance.runtime_properties
     if SSH_KEY not in host_rt_properties:
         host_rt_properties[SSH_KEY] = {}
-        host_rt_properties[SSH_KEY][PATH] = ctx.target.instance.runtime_properties[PRIVATE_KEY][PATH]
-        host_rt_properties[SSH_KEY][KEY] = ctx.target.instance.runtime_properties[PRIVATE_KEY][KEY]
-        host_rt_properties[SSH_KEY][USER] = ctx.target.instance.runtime_properties[PUBLIC_KEY][USER]
+        host_rt_properties[SSH_KEY][PATH] = target_rt_properties[PRIVATE_KEY].get(PATH)
+        host_rt_properties[SSH_KEY][KEY] = target_rt_properties[PRIVATE_KEY].get(KEY)
+        host_rt_properties[SSH_KEY][USER] = target_rt_properties[PUBLIC_KEY].get(USER)
 
 
 @operation
@@ -131,5 +132,7 @@ def _save_key_file(path, value):
         content_file.write(value)
 
 
-def _delete_key_file(path):
-    os.unlink(os.path.expanduser(path))
+def _delete_key_file(properties):
+    if PRIVATE_KEY in properties and PATH in properties[PRIVATE_KEY]:
+        path = properties[PRIVATE_KEY][PATH]
+        os.unlink(os.path.expanduser(path))
